refactor(emit): split static class expression helpers
