+++ crates/tsz-emitter/src/emitter/es5/helpers_async.rs
@@ -6,38 +6,13 @@
 
 use super::super::*;
 use super::helpers::ArraySegment;
+use super::helpers_class_expression_static::Es5StaticClassExpressionElement;
 use crate::emitter::core::PropertyNameEmit;
 use crate::emitter::declarations::class::replace_identifier;
 use crate::transforms::emit_utils;
 use crate::transforms::ir::IRNode;
 use crate::transforms::ir_printer::IRPrinter;
 use std::sync::Arc;
-
-#[derive(Clone)]
-pub(in crate::emitter) struct Es5StaticClassExpressionField {
-    pub name_emit: PropertyNameEmit,
-    pub initializer: NodeIndex,
-    pub member_pos: u32,
-}
-
-#[derive(Clone)]
-pub(in crate::emitter) enum Es5StaticClassExpressionElement {
-    Field(Es5StaticClassExpressionField),
-    StaticBlock {
-        block: NodeIndex,
-        saved_comment_idx: usize,
-        member_pos: u32,
-    },
-}
-
-impl Es5StaticClassExpressionElement {
-    pub(in crate::emitter) const fn member_pos(&self) -> u32 {
-        match self {
-            Self::Field(field) => field.member_pos,
-            Self::StaticBlock { member_pos, .. } => *member_pos,
-        }
-    }
-}
 
 impl<'a> Printer<'a> {
     fn next_arguments_capture_name(&mut self) -> String {
@@ -48,134 +23,6 @@ impl<'a> Printer<'a> {
                 return candidate;
             }
         }
-    }
-
-    pub(in crate::emitter) fn es5_static_class_expression_elements(
-        &self,
-        class_data: &tsz_parser::parser::node::ClassData,
-    ) -> Vec<Es5StaticClassExpressionElement> {
-        let mut inits = Vec::new();
-
-        for &member_idx in &class_data.members.nodes {
-            let Some(member_node) = self.arena.get(member_idx) else {
-                continue;
-            };
-            if member_node.kind == syntax_kind_ext::CLASS_STATIC_BLOCK_DECLARATION {
-                inits.push(Es5StaticClassExpressionElement::StaticBlock {
-                    block: member_idx,
-                    saved_comment_idx: self.static_block_inner_comment_index(member_node),
-                    member_pos: member_node.pos,
-                });
-                continue;
-            }
-            if member_node.kind != syntax_kind_ext::PROPERTY_DECLARATION {
-                continue;
-            }
-            let Some(prop) = self.arena.get_property_decl(member_node) else {
-                continue;
-            };
-            if prop.initializer.is_none()
-                || !self.has_effective_static_modifier_js(&prop.modifiers)
-                || self
-                    .arena
-                    .has_modifier(&prop.modifiers, SyntaxKind::AccessorKeyword)
-                || self
-                    .arena
-                    .has_modifier(&prop.modifiers, SyntaxKind::AbstractKeyword)
-                || self
-                    .arena
-                    .has_modifier(&prop.modifiers, SyntaxKind::DeclareKeyword)
-                || self
-                    .arena
-                    .get(prop.name)
-                    .is_some_and(|n| n.kind == SyntaxKind::PrivateIdentifier as u16)
-            {
-                continue;
-            }
-
-            if let Some(name_emit) = self.get_property_name_emit(prop.name) {
-                inits.push(Es5StaticClassExpressionElement::Field(
-                    Es5StaticClassExpressionField {
-                        name_emit,
-                        initializer: prop.initializer,
-                        member_pos: member_node.pos,
-                    },
-                ));
-            }
-        }
-
-        inits.sort_by_key(Es5StaticClassExpressionElement::member_pos);
-        inits
-    }
-
-    fn es5_class_expression_has_instance_field_where(
-        &self,
-        class_data: &tsz_parser::parser::node::ClassData,
-        name_filter: impl Fn(NodeIndex) -> bool,
-    ) -> bool {
-        class_data.members.nodes.iter().copied().any(|member_idx| {
-            let Some(member_node) = self.arena.get(member_idx) else {
-                return false;
-            };
-            if member_node.kind != syntax_kind_ext::PROPERTY_DECLARATION {
-                return false;
-            }
-            let Some(prop) = self.arena.get_property_decl(member_node) else {
-                return false;
-            };
-            if self.has_effective_static_modifier_js(&prop.modifiers)
-                || self
-                    .arena
-                    .has_modifier(&prop.modifiers, SyntaxKind::AccessorKeyword)
-                || self
-                    .arena
-                    .has_modifier(&prop.modifiers, SyntaxKind::AbstractKeyword)
-                || self
-                    .arena
-                    .has_modifier(&prop.modifiers, SyntaxKind::DeclareKeyword)
-            {
-                return false;
-            }
-            name_filter(prop.name)
-        })
-    }
-
-    fn es5_class_expression_has_instance_fields(
-        &self,
-        class_data: &tsz_parser::parser::node::ClassData,
-    ) -> bool {
-        self.es5_class_expression_has_instance_field_where(class_data, |_| true)
-    }
-
-    fn es5_class_expression_has_computed_instance_fields(
-        &self,
-        class_data: &tsz_parser::parser::node::ClassData,
-    ) -> bool {
-        self.es5_class_expression_has_instance_field_where(class_data, |name_idx| {
-            self.arena
-                .get(name_idx)
-                .is_some_and(|n| n.kind == syntax_kind_ext::COMPUTED_PROPERTY_NAME)
-        })
-    }
-
-    pub(in crate::emitter) fn static_block_inner_comment_index(&self, member_node: &Node) -> usize {
-        let brace_pos = if let Some(text) = self.source_text_for_map() {
-            let bytes = text.as_bytes();
-            let start = std::cmp::min(member_node.pos as usize, bytes.len());
-            let end = std::cmp::min(member_node.end as usize, bytes.len());
-            bytes[start..end]
-                .iter()
-                .position(|&byte| byte == b'{')
-                .map(|offset| (start + offset + 1) as u32)
-                .unwrap_or(member_node.end)
-        } else {
-            member_node.end
-        };
-        let mut idx = self.comment_emit_idx;
-        while idx < self.all_comments.len() && self.all_comments[idx].end <= brace_pos {
-            idx += 1;
-        }
-        idx
     }
 
     fn es5_class_iife_expression_from_var(output: &str, class_name: &str) -> Option<String> {
